Make nbsp prompt search visible and use b to cycle through results
Alacritty defaults to text: CellBackground and cursor: CellForeground
for the cursor, i.e. it inverts the character beneath. This doesn't play
nicely with wincent's trick to use a (nb) space to search for the last
command in tmux, because tmux highlights it, and then the cursor inverts
that, which means the net effect is nil, i.e. it looks like any other
space. This change sets the cursor explicitly to the standard iceberg
colors.

It also adds a shortcut b in copy mode to go back to the previous
prompt (i.e. prefix b to find the last command, then b, b, b, etc. to go
back succesive commands).