Make mouse/paste work better in regular Vim
This change removes the mapping of <esc> to nothing, which was breaking
pretty much everything involving the mouse and escape characters in
regular vim, including bracketed paste mode and italics.

Apparently it's ok to remap esc in neovim, but not in regular vim
(https://stackoverflow.com/a/8488537/409879), which is why the weirdness
was vim-only.

This commit also restructures the conditional configuration of regular
vim.