Add sd (size directory)
This won't work on Linux because gfind/gsed aren't there. I could fix
that with PATH munging or tests, etc. but ...

I should make this function more portable, e.g. the find can be replaced
with https://unix.stackexchange.com/a/4857/20079

And the sed can use escape sequences like this
https://unix.stackexchange.com/a/45954/20079