Fix set +x bug
Previously this script disabled set -x before running the eval (to
prevent the token appearing in stdout) and then enabled set -x after the
eval.

This means it left the environment with set -x regardless of what it was
before the script ran.

This change fixes that bug. It uses $-
(https://www.gnu.org/software/bash/manual/html_node/The-Set-Builtin.html)
to save the current set of options before disabling set -x, then it
restores the options.